fix: repair precedence regressions and complete strict-gate sweep
The earlier removal pass used textual replacement and broke grouping
on expressions whose redundant wrapper also provided parentheses:
str(x or '').strip() became x or ‘’.strip() (stripping the empty
string instead of the fallback), int(a or 0)/int(b or 0) became
a or 0/b or 0, and a token sum lost its grouping - caught by CI unit
tests. All such sites are restored with explicit parentheses.

Upstream meanwhile added the dev/pyrefly-check-local strict gate
(disabled ignore-files, broader coverage), which surfaces 231 more
instances than the default-config sweep did; those are removed as
well, following ruff's guidance for the ternary/comprehension
fallout. Gate exits clean locally; the previously failing suites
(workflow generator, agent observability) pass: 1576 passed.

+++ api/libs/token.py
@@ -116,7 +116,7 @@ def set_access_token_to_cookie(request: Request, response: Response, token: str,
         domain=_cookie_domain(),
         secure=is_secure(),
         samesite=samesite,
-        max_age=int(dify_config.ACCESS_TOKEN_EXPIRE_MINUTES * 60),
+        max_age=dify_config.ACCESS_TOKEN_EXPIRE_MINUTES * 60,
         path="/",
     )
 
@@ -142,7 +142,7 @@ def set_csrf_token_to_cookie(request: Request, response: Response, token: str):
         domain=_cookie_domain(),
         secure=is_secure(),
         samesite="Lax",
-        max_age=int(60 * dify_config.ACCESS_TOKEN_EXPIRE_MINUTES),
+        max_age=60 * dify_config.ACCESS_TOKEN_EXPIRE_MINUTES,
         path="/",
     )
